Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/referrer-policy/strict-origin-when-cross-origin/attr-referrer/cross-origin/http-http/script-tag/swap-origin-redirect/cross-insecure.http-expected.txt


FAIL The referrer URL is origin when a
                                 document served over http requires an http
                                 sub-resource via script-tag using the attr-referrer
                                 delivery method with swap-origin-redirect and when
                                 the target request is cross-origin. assert_equals: Reported Referrer URL is 'origin'. expected "http://localhost:8800/" but got "http://localhost:8800/referrer-policy/strict-origin-when-cross-origin/attr-referrer/cross-origin/http-http/script-tag/swap-origin-redirect/cross-insecure.http.html"
